+++ theme-css/src/commonMain/kotlin/tz/co/asoft/CSSTheme.kt
@@ -1,0 +1,3 @@
+package tz.co.asoft
+
+typealias CSSTheme = Theme<Typography>