+++ theme-css/src/commonMain/kotlin/tz/co/asoft/Typography.kt
@@ -2,7 +2,7 @@ package tz.co.asoft
 
 import kotlinx.css.rem
 
-class Typography(
+data class Typography(
     val h1: TextStyle = TextStyle(fontSize = 2.rem),
     val h2: TextStyle = TextStyle(fontSize = 1.5.rem),
     val h3: TextStyle = TextStyle(fontSize = 1.17.rem),