+++ theme-react/src/main/kotlin/tz/co/asoft/ThemeProvider.kt
@@ -9,6 +9,7 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import react.*
 
+@JsExport
 private class ThemeProvider(p: Props) : RComponent<ThemeProvider.Props, ThemeProvider.State>(p), CoroutineScope by CoroutineScope(SupervisorJob()) {
     class Props(
         val observerFrom: StateFlow<ReactTheme>,